refactor : 사용하지 않는 클래스 삭제 및 경로 수정
