+++ src/main/java/com/paintourcolor/odle/service/UserServiceInterface.java
@@ -1,7 +1,7 @@
 package com.paintourcolor.odle.service;
 
-import com.paintourcolor.odle.dto.TokenDto;
-import com.paintourcolor.odle.dto.TokenRequest;
+import com.paintourcolor.odle.dto.jwt.TokenResponse;
+import com.paintourcolor.odle.dto.jwt.TokenRequest;
 import com.paintourcolor.odle.dto.user.request.UserInactivateRequest;
 import com.paintourcolor.odle.dto.user.request.UserLoginRequest;
 import com.paintourcolor.odle.dto.user.request.UserSignupRequest;
@@ -11,8 +11,8 @@ import javax.servlet.http.HttpServletResponse;
 
 public interface UserServiceInterface {
     void signupUser(UserSignupRequest userSignupRequest);
-    TokenDto loginUser(UserLoginRequest userLoginRequest, HttpServletResponse response);
-    TokenDto reissue(TokenRequest tokenRequest);
+    TokenResponse loginUser(UserLoginRequest userLoginRequest, HttpServletResponse response);
+    TokenResponse reissue(TokenRequest tokenRequest);
     void logoutUser(String token);  // ?
     void inactivateUser(String username, UserInactivateRequest userInactivateRequest); // 이거 리퀘스트 같이 써도 되는지,,,
     UserResponse getUser(int page);
