--- src/main/java/com/paintourcolor/odle/util/jwtutil/JwtFilter.java
@@ -1,55 +1,0 @@
-//package com.paintourcolor.odle.util.jwtutil;
-//
-//import com.paintourcolor.odle.repository.LogoutTokenRepository;
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.security.core.Authentication;
-//import org.springframework.security.core.context.SecurityContextHolder;
-//import org.springframework.util.StringUtils;
-//import org.springframework.web.filter.OncePerRequestFilter;
-//
-//import javax.servlet.FilterChain;
-//import javax.servlet.ServletException;
-//import javax.servlet.http.HttpServletRequest;
-//import javax.servlet.http.HttpServletResponse;
-//import java.io.IOException;
-//
-//@RequiredArgsConstructor
-//public class JwtFilter extends OncePerRequestFilter {
-//    public static final String AUTHORIZATION_HEADER = "Authorization";
-//    public static final String BEARER_PREFIX = "Bearer ";
-//
-////    private final TokenProvider tokenProvider;
-//    private final JwtUtil jwtUtil;
-////    private final LogoutTokenRepository logoutTokenRepository;
-//
-//    // 실제 필터링 로직은 doFilterInternal 에 들어감
-//    // JWT 토큰의 인증 정보를 현재 쓰레드의 SecurityContext 에 저장하는 역할 수행
-//
-////    @Override
-////    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-////        // 1. Request Header 에서 토큰을 꺼냄
-////        String jwt = resolveToken(request);
-////
-////        // 2. validateToken 으로 토큰 유효성 검사
-////        // 정상 토큰이면 해당 토큰으로 Authentication 을 가져와서 SecurityContext 에 저장
-////        if (StringUtils.hasText(jwt) && jwtUtil.validateToken(jwt)) {
-////            Authentication authentication = jwtUtil.getAuthentication(jwt);
-////            SecurityContextHolder.getContext().setAuthentication(authentication);
-////        }
-//
-////        if(logoutTokenRepository.existsByToken(jwt)){
-////            throw new IllegalArgumentException("이미 로그아웃 처리된 토큰입니다");
-////        }
-//
-////        filterChain.doFilter(request, response);
-////    }
-//
-//    // Request Header 에서 토큰 정보를 꺼내오기
-////    private String resolveToken(HttpServletRequest request) {
-////        String bearerToken = request.getHeader(AUTHORIZATION_HEADER);
-////        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith(BEARER_PREFIX)) {
-////            return bearerToken.substring(7);
-////        }
-////        return null;
-////    }
-//}
