--- src/main/java/com/paintourcolor/odle/util/jwtutil/TokenProvider.java
@@ -1,116 +1,0 @@
-//package com.paintourcolor.odle.util.jwtutil;
-//
-//import com.paintourcolor.odle.dto.TokenDto;
-//import io.jsonwebtoken.*;
-//import io.jsonwebtoken.io.Decoders;
-//import io.jsonwebtoken.security.Keys;
-//import lombok.extern.slf4j.Slf4j;
-//import org.springframework.beans.factory.annotation.Value;
-//import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-//import org.springframework.security.core.Authentication;
-//import org.springframework.security.core.GrantedAuthority;
-//import org.springframework.security.core.authority.SimpleGrantedAuthority;
-//import org.springframework.security.core.userdetails.User;
-//import org.springframework.security.core.userdetails.UserDetails;
-//import org.springframework.stereotype.Component;
-//
-//import java.security.Key;
-//import java.util.Arrays;
-//import java.util.Collection;
-//import java.util.Date;
-//import java.util.stream.Collectors;
-//
-//@Slf4j
-//@Component
-//public class TokenProvider {
-//    public static final String AUTHORIZATION_HEADER = "Authorization_accessToken";
-//    public static final String REFRESH_AUTHORIZATION_HEADER = "Authorization_refreshToken";
-//    private static final String ROLE_KEY = "role";
-//    private static final String BEARER_TYPE = "Bearer";
-//    private static final String EMAIL_KEY = "email";
-//    private static final Long ACCESS_TOKEN_EXPIRE_TIME = (long) (1000 * 60 * 30); // 30분
-//    private static final Long REFRESH_TOKEN_EXPIRE_TIME = (long) (1000 * 60 * 60 * 24 * 7); // 7일
-//
-//    private final Key key;
-//
-//    public TokenProvider(@Value("${jwt.secret.key}") String secretKey) {
-//        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
-//        this.key = Keys.hmacShaKeyFor(keyBytes);
-//    }
-//
-////    public TokenDto generateTokenDto(Authentication authentication) {
-////        // 권한 가져오기
-////        String authorities = authentication.getAuthorities().stream()
-////                .map(GrantedAuthority::getAuthority)
-////                .collect(Collectors.joining(","));
-////
-////        long now = (new Date()).getTime();
-////
-////        // Access Token 생성
-////        Date accessTokenExpiresIn = new Date(now + ACCESS_TOKEN_EXPIRE_TIME);
-////        String accessToken = Jwts.builder()
-//////                .setSubject(authentication.getName())       // payload "sub" : "email"
-////                .claim(EMAIL_KEY, authentication.getName())   // payload "email" : "email"
-////                .claim(ROLE_KEY, authorities)        // payload "auth" : "USER"
-////                .setExpiration(accessTokenExpiresIn)        // payload "exp" : 1516239022
-////                .signWith(key, SignatureAlgorithm.HS512)    // header "alg" : HS512
-////                .compact();
-////
-////        // Refresh Token 생성
-////        String refreshToken = Jwts.builder()
-////                .setExpiration(new Date(now + REFRESH_TOKEN_EXPIRE_TIME))
-////                .signWith(key, SignatureAlgorithm.HS512)
-////                .compact();
-////
-////        return TokenDto.builder()
-////                .grantType(BEARER_TYPE)
-////                .accessToken(accessToken)
-////                .accessTokenExpiresIn(accessTokenExpiresIn.getTime())
-////                .refreshToken(refreshToken)
-////                .build();
-////    }
-//
-////    public Authentication getAuthentication(String accessToken) {
-////        // 토큰 복호화
-////        Claims claims = parseClaims(accessToken);
-////
-////        if (claims.get(ROLE_KEY) == null) {
-////            throw new RuntimeException("권한 정보가 없는 토큰입니다.");
-////        }
-////
-////        // 클레임에서 권한 정보 가져오기
-////        Collection<? extends GrantedAuthority> authorities =
-////                Arrays.stream(claims.get(ROLE_KEY).toString().split(","))
-////                        .map(SimpleGrantedAuthority::new)
-////                        .collect(Collectors.toList());
-////
-////        // UserDetails 객체를 만들어서 Authentication 리턴
-////        UserDetails principal = new User(claims.getSubject(), "", authorities);
-////
-////        return new UsernamePasswordAuthenticationToken(principal, "", authorities);
-////    }
-//
-//    public boolean validateToken(String token) {
-//        try {
-//            Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-//            return true;
-//        } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
-//            log.info("잘못된 JWT 서명입니다.");
-//        } catch (ExpiredJwtException e) {
-//            log.info("만료된 JWT 토큰입니다.");
-//        } catch (UnsupportedJwtException e) {
-//            log.info("지원되지 않는 JWT 토큰입니다.");
-//        } catch (IllegalArgumentException e) {
-//            log.info("JWT 토큰이 잘못되었습니다.");
-//        }
-//        return false;
-//    }
-//
-//    private Claims parseClaims(String accessToken) {
-//        try {
-//            return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(accessToken).getBody();
-//        } catch (ExpiredJwtException e) {
-//            return e.getClaims();
-//        }
-//    }
-//}
